setup/requirements.txt
pysha3==1.0.2
coincurve==18.0.0
watchdog==2.3.1
eth-keys==0.4.0
eth-utils==2.2.0
